pyproject.toml
[tool.poetry]
name = "simple_fastapi"
version = "0.1.0"
description = "Basic template with fastapi inside and other needed tools"
authors = ["Nikolay Baryshnikov <root@k0d.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.3.4"
ipython = "^7.11.1"
aioredis = "^1.3.1"
pydantic = "^1.4"
uvicorn = "^0.11.3"
fastapi = "^0.54.0"
requests = "^2.23.0"
aiofiles = "^0.4.0"
pytest-asyncio = "^0.10.0"
motor = "^2.1.0"
tornado = "^6.0.4"
aio-pika = "^6.6.0"
jinja2 = "^2.11.1"
pytils = "^0.3"
pytz = "^2019.3"
Pillow = "^7.1.1"
lxml = "^4.5.0"
gino_starlette = "^0.1.1"
gino = {version = "^1.0rc4", extras = ["gino_starlette"], allow-prereleases = true}
alembic = "^1.4.2"
sqlalchemy-utils = "^0.36.3"
psycopg2-binary = "^2.8.5"
starlette-prometheus = "^0.6.0"
sentry_sdk = "^0.14.3"
pydantic-sqlalchemy = "^0.0.4"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
pytest = "^5.3.5"
flake8 = "^3.7.9"
pytest-asyncio = "^0.10.0"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
